+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/models/export/NcbiSubmissionModel.java
@@ -21,9 +21,9 @@ public class NcbiSubmissionModel {
 	private final String bioProject;
 	private final String ncbiNamespace;
 	private final Date releaseDate;
-	private final List<NcbiBioSampleFilesModel> bioSampleFiles;
+	private final List<NcbiBioSampleModel> bioSamples;
 
-	public NcbiSubmissionModel(NcbiExportSubmission submission) {
+	public NcbiSubmissionModel(NcbiExportSubmission submission, List<NcbiBioSampleModel> bioSamples) {
 		this.id = submission.getId();
 		this.project = new ProjectMinimalModel(submission.getProject());
 		this.state = submission.getUploadState().name();
@@ -33,7 +33,7 @@ public class NcbiSubmissionModel {
 		this.bioProject = submission.getBioProjectId();
 		this.ncbiNamespace = submission.getNcbiNamespace();
 		this.releaseDate = submission.getReleaseDate();
-		this.bioSampleFiles = submission.getBioSampleFiles().stream().map(NcbiBioSampleFilesModel::new).collect(Collectors.toList());
+		this.bioSamples = bioSamples;
 	}
 
 	public Long getId() {
@@ -72,7 +72,7 @@ public class NcbiSubmissionModel {
 		return releaseDate;
 	}
 
-	public List<NcbiBioSampleFilesModel> getBioSampleFiles() {
-		return bioSampleFiles;
+	public List<NcbiBioSampleModel> getBioSamples() {
+		return bioSamples;
 	}
 }