+++ src/main/java/ca/corefacility/bioinformatics/irida/ria/web/services/UISampleService.java
@@ -11,7 +11,6 @@ import java.io.OutputStreamWriter;
 import java.nio.file.attribute.BasicFileAttributes;
 
 import java.util.*;
-import java.util.concurrent.atomic.AtomicReference;
 import java.util.stream.Collectors;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipOutputStream;
@@ -76,7 +75,10 @@ import ca.corefacility.bioinformatics.irida.ria.web.samples.dto.ShareSamplesRequ
 import ca.corefacility.bioinformatics.irida.repositories.specification.ProjectSampleJoinSpecification;
 import ca.corefacility.bioinformatics.irida.repositories.specification.SearchCriteria;
 import ca.corefacility.bioinformatics.irida.repositories.specification.SearchOperation;
+import ca.corefacility.bioinformatics.irida.ria.web.ajax.dto.SampleFilesResponse;
 import ca.corefacility.bioinformatics.irida.ria.web.exceptions.UIShareSamplesException;
+import ca.corefacility.bioinformatics.irida.ria.web.models.sequenceFile.PairedEndSequenceFileModel;
+import ca.corefacility.bioinformatics.irida.ria.web.models.sequenceFile.SingleEndSequenceFileModel;
 import ca.corefacility.bioinformatics.irida.ria.web.models.tables.AntSearch;
 import ca.corefacility.bioinformatics.irida.ria.web.models.tables.AntTableResponse;
 import ca.corefacility.bioinformatics.irida.ria.web.projects.dto.ProjectCartSample;
@@ -126,9 +128,9 @@ public class UISampleService {
 	 These correspond to their internationalized strings in the messages file
 	 */
 	private final List<String> TABLE_HEADERS = ImmutableList.of("server.SamplesTable.sampleName",
-			"server.SamplesTable.sampleId", "server.SamplesTable.quality", "server.SamplesTable.organism",
-			"server.SamplesTable.project", "server.SamplesTable.projectId", "server.SamplesTable.collectedBy",
-			"server.SamplesTable.created", "server.SamplesTable.modified");
+			"server.SamplesTable.sampleId", "server.SamplesTable.quality", "server.SamplesTable.coverage",
+			"server.SamplesTable.organism", "server.SamplesTable.project", "server.SamplesTable.projectId",
+			"server.SamplesTable.collectedBy", "server.SamplesTable.created", "server.SamplesTable.modified");
 
 	@Autowired
 	public UISampleService(SampleService sampleService, ProjectService projectService,
@@ -552,6 +554,28 @@ public class UISampleService {
 		return new SampleFiles(singles, filePairs, fast5, genomeAssemblies);
 	}
 
+	public SampleExportFiles getSampleExportFiles(Long sampleId, Long projectId) {
+		Sample sample = sampleService.read(sampleId);
+		// get the project if available
+		Project project = null;
+		if (projectId != null) {
+			project = projectService.read(projectId);
+		}
+
+
+		List<PairedEndSequenceFileModel> filePairs = getPairedSequenceFilesForExportSample(sample, project).stream()
+				.map(pair -> new PairedEndSequenceFileModel((SequenceFilePair) pair))
+				.collect(Collectors.toList());
+		List<SingleEndSequenceFileModel> singles = getSingleEndSequenceFilesForExportSample(sample, project).stream()
+				.map(single -> new SingleEndSequenceFileModel((SingleEndSequenceFile) single))
+				.collect(Collectors.toList());
+		List<SequencingObject> fast5 = getFast5FilesForExportSample(sample);
+		List<GenomeAssembly> genomeAssemblies = getGenomeAssembliesForExportSample(sample);
+
+
+		return new SampleExportFiles(singles, filePairs, fast5, genomeAssemblies);
+	}
+
 	/**
 	 * Get updated sample sequencing objects for given sequencing object ids
 	 *
@@ -654,6 +678,19 @@ public class UISampleService {
 		response.setHeader("Content-Disposition", "attachment; filename=\"" + genomeAssembly.getLabel() + "\"");
 		Files.copy(path, response.getOutputStream());
 		response.flushBuffer();
+	}
+
+	/**
+	 * Get details about the files belonging to a list of samples
+	 *
+	 * @param sampleIds - List of sample identifiers to get file details for
+	 * @param projectId - the project id that these samples belong to
+	 * @return A map of sample id and their related file information
+	 */
+	public SampleFilesResponse getFilesForSamples(List<Long> sampleIds, Long projectId) {
+		SampleFilesResponse response = new SampleFilesResponse();
+		sampleIds.stream().forEach(id -> response.put(id, getSampleExportFiles(id, projectId)));
+		return response;
 	}
 
 	/**
@@ -770,7 +807,7 @@ public class UISampleService {
 	 * @param obj     the {@link SequencingObject} to enhance
 	 * @param project the {@link Project} to add
 	 */
-	private void enhanceQcEntries(SequencingObject obj, Project project) {
+	public void enhanceQcEntries(SequencingObject obj, Project project) {
 		Set<QCEntry> availableEntries = new HashSet<>();
 		if (obj.getQcEntries() != null) {
 			for (QCEntry q : obj.getQcEntries()) {
@@ -785,32 +822,6 @@ public class UISampleService {
 	}
 
 	/**
-	 * Get a list of all {@link Sample} identifiers within a specific project
-	 *
-	 * @param projectId Identifier for a {@link Project}
-	 * @return {@link List} of {@link Sample} identifiers
-	 */
-	public List<Long> getSampleIdsForProject(Long projectId) {
-		Project project = projectService.read(projectId);
-		List<Sample> samples = sampleService.getSamplesForProjectShallow(project);
-		return samples.stream().map(Sample::getId).collect(Collectors.toUnmodifiableList());
-	}
-
-	/**
-	 * Get a list of all {@link Sample} names within a specific project
-	 *
-	 * @param projectId Identifier for a {@link Project}
-	 * @return {@link List} of {@link Sample} names
-	 */
-	public List<String> getSampleNamesForProject(Long projectId) {
-		Project project = projectService.read(projectId);
-		List<Sample> samples = sampleService.getSamplesForProjectShallow(project);
-		return samples.stream().map(Sample::getLabel).collect(Collectors.toUnmodifiableList());
-	}
-
-	/**
-	 * Share / Move samples with another project
-	 *
 	 * @param request Request containing the details of the move
 	 * @param locale  current users {@link Locale}
 	 * @throws UIShareSamplesException if project or samples cannot be found
@@ -1345,27 +1356,126 @@ public class UISampleService {
 	}
 
 	/**
+	 * Get a list of paired end sequence files for a sample
+	 *
+	 * @param sample  the {@link Sample} to get the files for.
+	 * @param project the {@link Project} the sample belongs to
+	 * @return list of paired end sequence files
+	 */
+	public List<SequencingObject> getPairedSequenceFilesForExportSample(Sample sample, Project project) {
+		Collection<SampleSequencingObjectJoin> filePairJoins = sequencingObjectService.getSequencesForSampleOfType(
+				sample, SequenceFilePair.class);
+		// add project to qc entries and filter any unavailable entries
+		List<SequencingObject> filePairs = new ArrayList<>();
+		for (SampleSequencingObjectJoin join : filePairJoins) {
+			SequencingObject obj = join.getObject();
+			enhanceQcEntries(obj, project);
+			filePairs.add(obj);
+		}
+
+		return filePairs;
+	}
+
+	/**
+	 * Get a list of single end sequence files for a sample
+	 *
+	 * @param sample  the {@link Sample} to get the files for.
+	 * @param project the {@link Project} the sample belongs to
+	 * @return list of single end sequence files
+	 */
+	public List<SequencingObject> getSingleEndSequenceFilesForExportSample(Sample sample, Project project) {
+		Collection<SampleSequencingObjectJoin> singleFileJoins = sequencingObjectService.getSequencesForSampleOfType(
+				sample, SingleEndSequenceFile.class);
+
+		List<SequencingObject> singles = new ArrayList<>();
+		for (SampleSequencingObjectJoin join : singleFileJoins) {
+			SequencingObject obj = join.getObject();
+			enhanceQcEntries(obj, project);
+			singles.add(obj);
+		}
+
+		return singles;
+	}
+
+	/**
+	 * Get a list of fast5 sequence files for a sample
+	 *
+	 * @param sample the {@link Sample} to get the files for.
+	 * @return list of fast5 sequence files
+	 */
+	public List<SequencingObject> getFast5FilesForExportSample(Sample sample) {
+		Collection<SampleSequencingObjectJoin> fast5FileJoins = sequencingObjectService.getSequencesForSampleOfType(
+				sample, Fast5Object.class);
+		return fast5FileJoins.stream()
+				.map(SampleSequencingObjectJoin::getObject)
+				.collect(Collectors.toList());
+	}
+
+	/**
+	 * Get any genome assemblies that are available for a sample
+	 *
+	 * @param sample the {@link Sample} to get the assemblies for
+	 * @return a list of genome assembly files
+	 */
+	public List<GenomeAssembly> getGenomeAssembliesForExportSample(Sample sample) {
+		Collection<SampleGenomeAssemblyJoin> genomeAssemblyJoins = genomeAssemblyService.getAssembliesForSample(sample);
+
+		return genomeAssemblyJoins.stream()
+				.map(SampleGenomeAssemblyJoin::getObject)
+				.collect(Collectors.toList());
+	}
+
+	/**
 	 * Format a Page of ProjectSampleJoin's into a format to be consumed by the Ant Design Table.
 	 *
 	 * @param page Page of {@link ProjectSampleJoin}
 	 * @return List of {@link ProjectSampleTableItem}
 	 */
 	private List<ProjectSampleTableItem> formatSamplesForTable(Page<ProjectSampleJoin> page, Locale locale) {
+		Map<Project, List<Long>> projectSampleIdsMap = page.getContent()
+				.stream()
+				.collect(Collectors.groupingBy(ProjectSampleJoin::getSubject,
+						Collectors.mapping(join -> join.getObject().getId(), Collectors.toList())));
+
+		Map<Project, Map<Long, Long>> projectSamplesCoverageMap = new HashMap<Project, Map<Long, Long>>();
+		projectSampleIdsMap.forEach((project, sampleIds) -> projectSamplesCoverageMap.put(project,
+				sampleService.getCoverageForSamplesInProject(project, sampleIds)));
+
+		List<Sample> samples = page.getContent().stream().map(join -> join.getObject()).collect(Collectors.toList());
+		Map<Long, List<QCEntry>> sampleQCEntries = sampleService.getQCEntriesForSamples(samples);
+
 		return page.getContent().stream().map(join -> {
 			Sample sample = join.getObject();
 			Project project = join.getSubject();
 
-			List<QCEntry> qcEntriesForSample = sampleService.getQCEntriesForSample(sample);
-			List<String> quality = new ArrayList<>();
+			Long coverage = null;
+			if (projectSamplesCoverageMap.containsKey(project)
+					&& projectSamplesCoverageMap.get(project).containsKey(sample.getId())) {
+				coverage = projectSamplesCoverageMap.get(project).get(sample.getId());
+			}
 
-			qcEntriesForSample.forEach(entry -> {
-				entry.addProjectSettings(project);
-				if (entry.getStatus() == QCEntry.QCEntryStatus.NEGATIVE) {
-					quality.add(messageSource.getMessage("sample.files.qc." + entry.getType(),
-							new Object[] { entry.getMessage() }, locale));
+			List<QCEntry> qcEntriesForSample = sampleQCEntries.get(sample.getId());
+			List<String> quality = new ArrayList<>();
+			String qcStatus = null;
+
+			// If the sample has any SequencingObjects we will have at minimum CoverageQCEntry's
+			// which can be checked to set QCStatus.
+			if (qcEntriesForSample != null) {
+				qcEntriesForSample.forEach(entry -> {
+					entry.addProjectSettings(project);
+					if (entry.getStatus() == QCEntry.QCEntryStatus.NEGATIVE) {
+						quality.add(messageSource.getMessage("sample.files.qc." + entry.getType(),
+								new Object[] { entry.getMessage() }, locale));
+					}
+				});
+				// set qcStatus based on filtered qcEntries
+				if (quality.size() == 0) {
+					qcStatus = "pass";
+				} else {
+					qcStatus = "fail";
 				}
-			});
-			return new ProjectSampleTableItem(join, quality);
+			}
+			return new ProjectSampleTableItem(join, quality, qcStatus, coverage);
 		}).collect(Collectors.toList());
 	}
 
@@ -1414,6 +1524,11 @@ public class UISampleService {
 			Cell sampleQualityCell = row.createCell(cellNum++);
 			sampleQualityCell.setCellValue(StringUtils.join(item.getQuality(), "; "));
 
+			Cell sampleCoverageCell = row.createCell(cellNum++);
+			if (item.getCoverage() != null) {
+				sampleCoverageCell.setCellValue(item.getCoverage());
+			}
+
 			Cell sampleOrganismCell = row.createCell(cellNum++);
 			sampleOrganismCell.setCellValue(sample.getOrganism());
 
@@ -1447,8 +1562,7 @@ public class UISampleService {
 	 *
 	 * @param response {@link HttpServletResponse}
 	 * @param filename {@link String} name of the file to download.
-	 * @param items    {@link ProjectSampleTableItem} details about each row of the table   Data to download in the
-	 *                 table
+	 * @param items    {@link ProjectSampleTableItem} details about each row of the table Data to download in the table
 	 * @param headers  for the table
 	 * @throws IOException thrown if file cannot be written
 	 */
@@ -1462,9 +1576,16 @@ public class UISampleService {
 			SampleObject sample = item.getSample();
 			ProjectObject project = item.getProject();
 			String[] row = {
-					sample.getSampleName(), sample.getId().toString(), StringUtils.join(item.getQuality(), "; "),
-					sample.getOrganism(), project.getName(), project.getId().toString(), sample.getCollectedBy(),
-					sample.getCreatedDate().toString(), sample.getModifiedDate().toString() };
+					sample.getSampleName(),
+					sample.getId().toString(),
+					StringUtils.join(item.getQuality(), "; "),
+					item.getCoverage() != null ? item.getCoverage().toString() : "",
+					sample.getOrganism(),
+					project.getName(),
+					project.getId().toString(),
+					sample.getCollectedBy(),
+					sample.getCreatedDate().toString(),
+					sample.getModifiedDate().toString() };
 			results.add(row);
 		}
 
@@ -1475,37 +1596,5 @@ public class UISampleService {
 		csvWriter.writeAll(results);
 		csvWriter.flush();
 		csvWriter.close();
-	}
-
-	/**
-	 * Check if a list of sample names exist within a project
-	 *
-	 * @param request Request containing the project id and sample names
-	 * @return List of valid and invalid sample names
-	 */
-	public SampleNameCheckResponse checkSampleNames(SampleNameCheckRequest request) {
-		Iterable<Project> projects = projectService.readMultiple(request.getProjectIds());
-		List<ValidSample> valid = new ArrayList<>();
-
-		AtomicReference<Iterator<Project>> iterator = new AtomicReference<>();
-		request.getNames().forEach(name -> {
-			Sample sample = null;
-			iterator.set(projects.iterator());
-
-			// Need to figure out what project it belongs to.
-			while (sample == null && iterator.get().hasNext()) {
-				Project project = iterator.get().next();
-				try {
-					sample = sampleService.getSampleBySampleName(project, name);
-					valid.add(new ValidSample(project, sample));
-				} catch (Exception e) {
-					// Nothing to worry about here
-				}
-			}
-		});
-		List<String> invalid = new ArrayList<>(request.getNames());
-		invalid.removeAll(valid.stream().map(ValidSample::getSampleName).collect(Collectors.toList()));
-
-		return new SampleNameCheckResponse(valid, invalid);
 	}
 }